requirements.txt
einops==0.7.0
ftfy==6.1.3
h5py==3.9.0
icecream==2.1.3
matplotlib==3.8.3
MedPy==0.4.0
monai==1.3.0
numpy==1.24.3
opencv_python_headless==4.9.0.80
Pillow==10.2.0
pycocotools==2.0.7
pynvml==11.5.0
PyYAML==6.0.1
regex==2023.12.25
safetensors==0.3.2
scipy==1.12.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2

